Attempt to deflake TetDiscovery (#35840)
Wait for the target to be registered before running any commands
against it. All of the most recent errors seem to be caused by
`direct dialing to nodes not found in inventory is not supported`.
By waiting for the node to be present this should not longer be a
possible failure scenario.