Add support for MFA in kube proxies in Connect (#35772)
* Improve error handling in gateway/kube_test.go

* Make gatewaytest.mustGenCertSignedWithCA public

It will be needed to generate and then return cert without saving it to
a file.

* Move AddMetadataToRetryableError to where it's actually needed

Cluster.reissueDBCerts is called in two scenarios:

* During the request to create a db gateway: this request is made by
  the Electron app, so we need to add metadata to let the app know that
  this error can be resolved by relogin.
* When someone attempts to use the proxy but the user cert has expired:
  this happens outside of the Electron app request-response cycle and
  thus adding metadata is unnecessary.

* Make gatewaytest.MustGenCACert public, accept CA for MustGenAndSaveCert

This will make writing tests where we need to supply either Cert or KeyPath
and CertPath a little bit easier.

* Kube gateways: Store cert in memory instead of on disk

* Add ctx to testGatewayCertRenewal

If a gateway requires MFA, daemon.Service.CreateGateway can potentially
block forever. To work around this, we want the tests to pass a context
around so that we can set a timeout on the context in MFA tests.

* Convert test function args into structs

This will make it easier to pass webauthnLogin for tests that need MFA.

* Refactor handling of mock tshd events service addr

* Add tests for MFA

* Do not save kube key in local agent

Connect issues a new cert for every gateway anyway.

* Check kube certs after issue in teleterm gateways

This makes sure that Connect does the same checks as tsh.

* reissueDBCerts → reissueGatewayCerts

* Pass mfaPrompt from daemon to GatewayCreator

This also fixes a problem where the mfaPrompt passed from GatewayCreator
was not the expected one. daemon.Service.ResolveClusterURI attaches
MFAPromptConstructor on clusterClient. However, GatewayCreator does not
call daemon.Service.ResolveClusterURI, it calls
clusters.Storage.ResolveCluster which does not attach the proper prompt
constructor.

* Refactor GetPromptReasonSessionMFA into WithPromptReasonSessionMFA

* Verify that PromptMFA on tshdEventsService gets called during MFA ceremony

* reissueKubeCert: Get rid of unused mfaRequired var

* Use fake storage in gateway CRUD tests

Now that daemon.Service.CreateGateway needs a TeleportClient, we have
to use a fake storage that's able to provide it.

* Fix synchronization between webauthnLogin and PromptMFA

* Use sync/atomic for function call counts