Allow fetch depth of checkout-action to be specified (#3017)
This is required to run the flaky test detector in teleport.e.
Without it the mechanism to determine which tests have been modified
would not work because it can't find the previous commit information.
The default behavior is preserved by making the input optional and
using the same default as the equivalent option in the actions/checkout.

Source-commit: gravitational/teleport.e@a10cc2745cc6a6573ef48eb31aec207ff6bcf6fc
Source-pull-requests: gravitational/teleport.e#3017