Fix downgrade logic of `KubernetesResources` to Role `v6` (#35991)
`KubernetesResources` were improperly downgraded when they grant access
to all resources.

In that case, the role was downgraded to something that can't be used to
access Kubernetes clusters but they could have been downgraded to a
`Role` v6 with the same permissions as the `KubernetesResources`.

This commit fixes the downgrade logic to downgrade to a `Role` v6 with
the same permissions as the `KubernetesResources`.

A role v7 with

```json
kubenretes_labels:
 '*': '*'
kubernetes_resources:
- kind: pod
  name: '*'
  namespace: '*'
  verbs:
  - '*'
```
Is downgraded to a role v6 with

```json
kubenretes_labels:
 '*': '*'
kubernetes_resources:
- kind: pod
  name: '*'
  namespace: '*'
```

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>