Add Machine ID and Kubernetes access diagrams
See #19190

This change adds architecture diagrams to the introduction pages of the
"Machine ID" and "Kubernetes Access" sections of the docs.

Making this change to the Kubernetes Access section presented an issue,
since there isn't an elegant way to include both the illustration video
that is currently on the page and the new architecture diagram. This
change replaces the video with the diagram since users tend not to spend
much time on the page.