Tests: check DiscoveryFetch events before creating resource (#36008)
After creating a DiscoveryConfig with valid matchers, the
DiscoveryService will fetch Cloud Resources and emit a
DiscoveryFetchEvent.

The test was asserting that the number of FetchEvents was zero right
after creating the DiscoveryConfig, but it might happen that between the
CreateDiscoveryConfig and reading the DiscoveryFetchEventCount, the
DiscoveryConfig was processed, a new fetch was triggered and an event
emitted.

This PR moves the check to before the CreateDiscoveryConfig call.