Copy FIPS build out of `e` on ARM64 (#3053)
FIPS builds end up in `e/build/artifacts`, as `$RELEASE_DIR` does not
passthrough to the docker build. This was already being done for AMD64,
so just needed to copy this over to ARM64 as well.

Also, `go` doesn't seem to be installed by default, so install it,
as we use it within `Makefile` multiple times.

Source-commit: gravitational/teleport.e@c11b84bb122954924be425e7a4cdae056b0e63be
Source-pull-requests: gravitational/teleport.e#3053