Add architecture diagrams to the docs (#36201)
Closes #19190

Ensure that all "\<Resource\> Access" sections of the docs include an
architecture diagram by adding diagrams to the "Server Access" and
"Application Access" sections.

The intro page of the Application Access section currently includes a
demo video. There is no elegant way to include both the demo video and
the architecture diagram. This change removes the demo video since it is
out of date.