Relax Kubernetes CRD discovery when building cache (#36214)
Teleport Kubernetes Service has a monitor that constantly watches the
Resources registered in the Kubernetes Cluster via API Discovery. The
goal is to keep an up-to-date representation of all resources existing
in the cluster in order to be able to register them for Teleport
per-Resource RBAC.

Having an up-to-date represenation allows us to unmarshal the API
responses and filter them when the custom resources are local.

When the Kubernetes APIs are registered using non-local services - i.e.
the API is served by a POD running within the cluster like metrics API -
and those services aren't healthy - i.e. pod not running, invalid
selector, cluster has no nodes - the discovery watcher returns an error
and fails. This is an improper configuration but seems to be a common
problem.

This PR relaxes the discovery mechanism and doesn't enforce that all
APIs return their resources if they aren't currently available.

When the `client.Discovery().ServerGroupsAndResources()` returns a
`*discovery.ErrGroupDiscoveryFailed`, it also returns the partial
results that we will use for registration.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>