Refactor hostname resolution for SSH connections via the WebUI (#35773)
* Refactor hostname resolution for SSH connections via the WebUI

Previously all hostname resolution was performed by sending either
the hostname or server uuid to Auth in a GetSSHTargets request.
The returned information was then sent to the UI so that the console
tab can be updated to include `user@host`. While this worked, it
required a round trip to Auth that had to iterate through the entire
set of nodes to do the resolution. During a concurrent session test
that tried to spawn several thousand web sessions it was discovered
that this cause a massive spike in CPU on Auth.

There are two reasons that Auth was used to resolve the hostname.
 1) The proxy doesn't perform RBAC on behalf of the user. Instead
    the proxy has an Auth client with the users identity.
 2) The UI used to provide an input box that users could manually
    enter the target host.

Since we no longer allow open dialing and the input box to enter
a connection string manually has been removed we no longer need
to worry about the second case. To avoid the round trip to auth
we can use the local Proxy cache to look up the hostname if we
wait until AFTER the ssh connection to the target is established.
At this point we can be sure that the user does have access to
the target since the node allowed the connection.

* Fix terminal tests

The testing utilities used to mimic the web ui half of the websocket
made assumptions that the session metadata would be the first message
provided. However, now that hostname resolution occurs after the
initial SSH connection has been established it is possible for other
messages to be sent first. This caused some tests to fail and others
to hang forever.

The duplicated terminal logic from (WebSuite) makeTerminal and
(testProxy makeTerminal has now consolidated into a single test
terminal utility. The new `terminal` now wraps a TerminalStream
which allows test code to make use of the message processing logic
already in place there instead of having to write custom logic
per test. All a test needs to do is provide handlers for any messages
that it desires to introspect.