Remove outdated Details box in Installation
Closes #33008

The changes in #36022 removed a `Details` box with outdated information
in the Installation page. This change backports that one. The
instructions in the `Details` box refer to installation steps that we
removed from the `install-linux.mdx` partial.

We will want to include more explicit instructions about using our DEB
package repo at some point, but for now, this change prevents possible
confusion.