fix: panic in external audit watcher
This commit fixes a possible panic in the cluster External Audit watcher
that can occur whenever the underlying backend watcher closes
unexpectedly, and resulted in a double-close of the `initialized`
channel.

This fix here is to switch from closing an "initialized" channel
whenever a watcher is initialized, to periodically writing to a
"running" channel during the events watch loop, whenever there is a
listener waiting on the channel.
This is what some other watchers (e.g. HeadlessAuthenticationWatcher)
already do, and allow for the test to wait for the new watcher to init
after simulating a failed watcher.

Changelog: Fixed potential panic after backend watcher failure.