Fix `rdp-client` FIPS builds (#36280)
* Fix `rdp-client` FIPS builds

gravitational/teleport.e#3072 fixed building rdp-client with FIPS, but uncovered other problems.

* Set `CMAKE=cmake3`, as `boring-rs` (by way of `cmake-rs`) calls `cmake` (which isn't installed).
* Upgrade `boring` and `tokio-boring` to v4.3.0 to pick up build fixes.

* Update Dockerfile-centos7 with LIBCLANG_PATH environment variable

This commit adds a necessary environment variable, LIBCLANG_PATH, to Dockerfile-centos7. This variable is needed for the boring-rs to generate bindings, as it points towards the location of libclang file.

* Add `golang` to node buildbox

The node buildbox was missing `go`, so it wasn't handling our makefiles correctly.

* Update Dockerfile-centos7 to install additional dev tools

This commit updates the build command in Dockerfile-centos7 to install 'clang-resource-headers' and 'libclang'. These additions are needed by boring-rs library to build correctly. We're building it as a part of our FIPS build.

---------

Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>