Validate the user still exists during account reset (#36271)
Prior to this, any reset links created before a user was deleted
would still be valid and when used would recreate the deleted user.
With this change, the user is validated to still exist prior to
honoring the reset request and if the user no longer exists any
remaining reset tokens are deleted.