Use GetRole in the user login state generator, fail if not found. (#36321)
The user login state generator will use GetRole instead of GetRoles to lookup
roles. GetRole has a more robust fallback when unable to find things in the
cache, so using it directly will make the user login state generator more
robust in general.

Additionally, the user login state generator should not allow missing roles,
so if a role is missing it will now throw an error.