feat: AWS KMS keystore backend (#36238)
* feat: AWS KMS keystore backend

This commit adds backend support for AWS KMS as a keystore backend for
Teleport CAs.
It is similar to the existing GCP KMS support.

This PR was getting a bit large, so it does not yet expose the
configuration parameters to enable this in a real cluster, which will
come in a following PR, along with docs.

* simplify string concat

* fix lint