Refactor desktop player
Adopt an approach similar to the SSH player (more standard react
calls and less event emitting). This also updates the progress
bar to be more of a "dumb" component that just renders state,
and leaves the smoothing out of progress updates to the client.

In addition, this adds support for seeking by dragging the
progress bar to a particular portion of the video.

Fixes #17199