[v14] Fix the access list reminder pagination loop. (#36407)
* Fix the access list reminder pagination loop.

On error in the access list reminder pagination loop, the loop now breaks
instead of continuing, which is the intended behavior. A test has been added
to exercise this. This test will time out with the previous behavior. A light
refactoring has been done to make it easier to mock the ListAccessLists
response.

Users running their own Slack plugins (outside of cloud) were running into an
issue here due to the plugin user not having permissions to read access lists.
On cloud tenants, the plugins should have this access, so the impact there is
minimal.

Additionally, the docs have been updated to include the access list
permissions.

* Test fixes.

* Tests are more consistent.

* Update docs/pages/includes/plugins/rbac-with-friendly-name.mdx

Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>

* Add in initial duration/trigger.

* Output migration instructions, add jitter back in.

* Don't return error when no recipients.

* Tune the error return.

---------

Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>