auth: Remove Role downgrade logic to v6
This PR removes the role downgrade logic introduced in Teleport 14 to
support Teleport 13 clients. This happens because v13 clients like
agents and tctl didn't support `v7` roles and it required custom
downgrade logic to reduce the access levels to a level where it was at
least as strickier as the rules defined in Role `v7`.

For Teleport 15, all clients will run at least version `v14` so we can
remove the custom downgrade logic.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>