Kubernetes: Cleanup code marked to be deleted for v15 (#36412)
This PR cleans code related to Kubernetes Proxies and Kubernetes
Services not supporting re-usage of the same underlying connection for
multiple users - aka proxy dials with its credentials and passes users
identities using Impersonation header.

The support for this feature was introduced in Teleport v13 and the old
code was scheduled to be deleted in Teleport v15.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>