feat: seamless migration to HSM/KMS
The PR makes it much easier to migrate an existing Teleport cluster from
software keys to HSM or KMS keys.

Previously, as soon as an HSM/KMS was configured in the teleport.yaml
for an auth server, it would immediately refuse to sign any more
certificates with software keys.
This was meant to defend against someone configuring an HSM and then
forgetting to perform the necessary CA migrations, thinking they were
protected by the HSM when in fact they were still using the old software
keys.

In practice, this just made it very difficult to migrate a cluster since
there would be downtime where you couldn't even use tctl remotely
because there was no way to log in.
In a dual-auth cluster you could theoretically avoid downtime but the
process was arcane and difficult to execute.
Check out the docs changes to see all the steps I was able to remove
here.

This will be critical for enabling Cloud to start using AWS KMS keys.

TODO: add a cluster alert (probably not for Cloud) when an HSM/KMS is
configured but not actively used yet because a CA rotation is needed.
Necessary now that `tctl status` no longer prints this (it has no way to
tell).