Update cloudflare/cfssl in order to fix enterprise builds
We now depend on cloudflare/cfssl transitively via
gravitational/license. The cfssl repo itself has a dependency on
zmap/zlint, which is outdated and doesn't build on newer versions
of Go:

    # github.com/zmap/zlint/util
    ../../../go/pkg/mod/github.com/zmap/zlint@v0.0.0-20190806154020-fd021b4cfbeb/util/encodings.go:38:24: cannot use atv.Type (variable of type "github.com/zmap/zcrypto/encoding/asn1".ObjectIdentifier) as "encoding/asn1".ObjectIdentifier value in argument to IsNameAttribute
    ../../../go/pkg/mod/github.com/zmap/zlint@v0.0.0-20190806154020-fd021b4cfbeb/util/oid.go:136:16: cannot use v.Type (variable of type "github.com/zmap/zcrypto/encoding/asn1".ObjectIdentifier) as "encoding/asn1".ObjectIdentifier value in argument to oid.Equal

The fix is to upgrade cfssl to a version that uses a newer version of
zlint. Cloud team resolved a similar issue in the past:
https://github.com/gravitational/cloud/pull/3805#discussion_r1143909550