Send the negotiated screen resolution to the browser (#36330)
* Send the negotiated screen resolution to the browser

With RDP the server doesn't always provide you with exactly the
resolution you asked for. When we only supported RDP6 bitmaps,
this was okay, the data we got back still "fit" and properly
rendered on a canvas even if the canvas was not quite sized
correctly.

With RemoteFX, the decoder needs to know the true screen resolution
in order to properly decode the image data. Our decoder runs in the
browser, which means we need to send some additional data to the
browser to indicate the resolution being used. It turns out, we already
send a "channel IDs" message at exactly the same time when we have
the screen resolution, so I've piggy backed on that message (and
renamed it for clarity).

Fixes #35994

* negotiated screen resolution polish (#36358)

---------

Co-authored-by: Isaiah Becker-Mayer <isaiah@goteleport.com>