Improve Linux installation instructions
Closes #32387

In the Installation page, the instructions for installing Teleport on
Linux reuses the partial we use for how-to guides, `install-linux.mdx`.
This is inappropriate for an installation reference, since the partial
is not intended for users interested in a specific package manager. This
change provides a comprehensive description of our package repositories
for Teleport installations on Linux.

Since the Installation page is a common entrypoint for users looking to
get started with Teleport, this change also adds a section to the Linux
installation instructions with recommended pages to follow for setting
up Teleport.

Also, since we have a script available form the corporate Teleport site
that automatically chooses a package manager and configures it to target
a Teleport package repo, this change adds instructions to use this.

I have organized the Linux installation section to place H3 sections
from less work/configuration to more work/configuration, from the
one-line installation script to package manager-specific instructions
and, finally, general information about our package repositories.