Fix goroutine leak in SSH service (#36512)
The session object was never closing the write half of the io.Pipe
that is created to propagate input of the session. This results in
the goroutine responsible for reading from the pipe to block
indefinitely. `(session) Stop` was updated to close the write half
of the pipe which causes the read goroutine to terminate properly
when the session is terminated.

Fixes #36487.