Always close completed app sessions (#36651)
* Extend expiry functionality of FnCache

The FnCacheConfig now provides an optional OnExpires callback which
will be called when any item in the cache is removed due to expiry.
Eviction of expired entries used to only happen on retrieval
attempts, but this can now be triggered on demand using
(FnCache) RemoveExpired. A new Shutdown method was also added to
allow all items in the cache to be removed. If OnExpires is set
then it will be called for every item in the cache during Shutdown.

* Add cancellation to abort in flight loads

* Always close completed app sessions

The previous use of gravitational/ttlmap with a fixed capacity led
to sessions leaking and never being closed if a spike of sessions
greater than the capcity occurred. The sessionChunkCache relied
on the expiry callback function set on the ttlmap to be called in
order to close the app session and release resources. However, if
an item was evicted from the ttlmap because the capacity had been
exceeded the expiry callback was fired.

Instead of increasing the capacity to a larger size the ttlmap was
replaced entirely by utils.FnCache. It offers equivalent
functionality while still being actively maintained. This also
allowed refactoring to remove the sessionChunkCache entirely since
the FnCache took care of most of the added functionality that the
ttlmap wrapper was providing.

Partially addressed #36541.