AWS OIDC DeployDatabaseServices - implement AutoUpgrades (#36695)
The new DeployService adds some non-determenistic logic with regards to
the deployed ECS Services.

Instead of looking for a specific ECS Service, we now iterate over all
the ECS Services of a well-known ECS Cluster (same as before) and update
the container image (as long as the tags are the Teleport ones).

Not everyone will have the new set of permissions which are created when
using the new flow.
For those situations, we keep the old behaviour where we know the
exact ECS Service name.