Fix DiscoveryFetch events test (#36797)
This PR fixes a test that has been failing for some days in the Daily
Flaky Test runner.

The producer either emits an event (iff the consumer is selecting for an
event).

The problem is that we were using an unbuffered channel, which means
that when the consumer was processing, no other notification was being
sent.

So, if the producer emitted an event, it would get consumed and a Fetch
would start. If the producer emitted another event, while the consumer
was still processing the event, it would not wait for it to finish and
the `default` branch of the `select` was selected and nothing happened.
So, when the consumer finished the Fetch operation it would wait either
for the next event or for the time-based-Poll to trigger a new event.

Adding a buffer to the channel, ensures that even if the consumer is
processing, a new event is added and a new Fetch starts right after the
Fetch operation finishes.