Prevent deadlock on moderated sessions when mod connection drops (#36919)
This PR removes a deadlock caused by the moderator leaving the session
because his connection drops.

`OnWriteError` was called under lock which creates an issue if the
function calls `Termnager.DeleteWriter` to exclude the writter from the
term manager.

This PR also correctly forwards errors to the clients when they occur.

Changelog: Ensure that moderated sessions do not get stuck in the event of an unexpected drop in the moderator's connection.

Fixes #36881

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>