docs: update license download page (#36673)
Going forward, self-hosted licenses will include an anonymization
key that is not known to Teleport. Customers who run multiple
Teleport clusters can use the same license file on each cluster,
and Teleport will de-duplicate users to ensure that usage is
tracked properly.

This anonymization key is generated each time a license is downloaded,
so it's important to download the license once and reuse it rather
than performing multiple downloads.