Remove misleading diagram (#36656)
Closes #35337

The diagram in the "Getting Started" page of the "Application Access"
section has been misleading to users, as it does not indicate which
Teleport components are involved in application routing.

There is now a diagram in the Introduction page of the section that is
clearer in terms of which components are involved in Teleport
application access and how they interact. This change removes the
diagram in the "Getting Started" page.

Note that the Introduction page diagram could use some improvement--it's
pretty overwhelming--but since users who put in the effort can still use
it to identify interactions between Teleport components, we can clarify
the diagram in a future change rather than in this one.