UI access list modification preserves metadata. (#3169)
Metadata for an access list is preserved when attempting to edit it through
the API. Right now, if an owner tries to modify membership for an access list
that has labels set, the owner will receive a message that they cannot modify
the access list. This is because the access list UI endpoint doesn't preserve
the labels, resulting in a technically different access list despite the user
not intending to do so.

By preserving the metadata, we can avoid this issue.

Source-commit: gravitational/teleport.e@2e4de3583a551c47e59c967c5679324ddca3bfa8
Source-pull-requests: gravitational/teleport.e#3169