[v15] Polishing desktop player (#37043)
* Waits for the canvas to be fully defined before rendering the canvas and
progress bar.

This prevents visual glitches at pageload where the canvas starts out small and
then suddenly expands to its full size (moving the progress bar down with it).

* Ensure play speed selector is disabled when playback completes

* Fixes speed selector focus issue

Because `clientOnClientScreenSpec` was not memoized, TdpClientCanvas
was constantly re-rendering, and some logic in it was causing the canvas
to steal focus.

This commit memoizes `clientOnClientScreenSpec`, and as a safeguard it
also wraps the focus stealing in a `useEffect` s.t. it only happens on
mount.

* Fixes https://github.com/gravitational/teleport/issues/36827 and polishes desktop playback error display

* use existing isComplete variable

* updating snapshot

---------

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>