[v15] Fix top bar icons (#37057)
* Increases the contrast between the enabled and disabled states
of the icons in the top bar and warning dropdown.

* Change from using title to using HoverTooltip in the DesktopSession's TopBar

* Create more accurate synchronization of clipboard sharing state

Corners had been cut in the past to make the clipboard sharing state
easier to manage, after the complexity of an original implemntation had
grown overly burdensome. This commit fixes those corners without wading
into nearly as much complexity as the original implementation.

* Updating snapshot