[v12] fix key generation for dual auths sharing a single YubiHSM2
Backport #36899 to branch/v12

The actual fix is a few characters in lib/auth/keystore/pkcs11.go.
I'm also backporting changes to test files from #36549 that this PR built on
top of, which make it easier to run all HSM unit and integration tests
with a connected YubiHSM2 (which I did when putting together this
backport).

Instead of merging all changes in the integration tests, I just checked
out the state of them from branch/v13 in
https://github.com/gravitational/teleport/pull/37301

Changelog: fixes CA key generation when two auth servers share a single YubiHSM2