[v15] docs: Fix description of AMI names with <arch> addition. (#36998)
* docs: Fix description of AMI names with <arch> addition.

Fix the names and descriptions of the AMI names now that we have two
architectures (x86_64 and arm64) for AMIs. The `<arch>` description was
added previously but was not added to the example, and the description
of the `-fips` part became incorrect or ambiguous.

* docs: Fix description of location of -fips suffix on AMIs (#36999)

* docs: Fix description of location of -fips suffix on AMIs

The `-fips` suffix appears after the `<arch>` component, not before as I
previously documented in 1548a43 (I saw some v15 dev builds with -fips
after the version, but they were from before <arch> was added).

This can be seen with:

    aws --region us-west-2 ec2 describe-images --owners 146628656107 --filters 'Name=name,Values=teleport-ent-15.0.0-alpha.5-*-fips-*'

* docs: Add example for FIPS AMI