integrations/operator: support multiple versions of the same resource in crdgen (#36715)
The commit contains the following changes:

- add multi-version support by creating new kinds. This is currently done via an override to maintain compatibility. At some point, we might migrate unversioned kinds (TeleportRole) to versioned kinds (TeleportRoleV5) and remove the override
- unify all resource statuses in a single struct. This simplifies the CRD building because we are not trying to guess which package is correct anymore when parsing go's AST
- update crdgen snapshots

This PR does not update the base CRDs in the operator or the Hlem chart. This will happen in the next PR, adding go types and controller support.

Note for reviewers: the changeset might seem large, but the changes are split into commits that are much more reviewable. You only need to review the first two commits.