Add service definition for Okta plugin configuration helper.
The gRPC service definition for the Okta plugin configuration helper will
allow users to query Okta directly using Okta credentials during plugin
configuration. This will then be used to craft filters and regexes that
the user may want to limit the importing to a subset.

OktaGroups and OktaApps are separate messages because there is potential
to add more data to these for display in the UI, though currently it only
houses the name.