[v15] dronegen: Remove operator OCI pipeline dependencies (#37147)
* dronegen: Remove operator OCI pipeline dependencies

Remove the dependencies from the pipeline for building the operator OCI
images. The GitHub workflow called used to build both the legacy OCI
images and the operator OCI images and the dependencies were only for
the legacy images. The called workflow no longer builds the legacy
images so these dependencies are no longer needed.

* drone: Regenerate .drone.yml for remove OCI deps

Regenerate .drone.yml with `make dronegen` to remove the pipeline
dependencies of the build-oci pipeline.