[v15] Add ninja-build setup to Dockerfile-centos7 (#37122)
* Add ninja-build setup to Dockerfile-centos7

The Dockerfile-centos7 has been updated to include the setup and installation of ninja-build. The version of ninja-build included in CentOS 7 AMR64 was too old for building boringssl, hence it needs to be built from source. Both additional dependencies and the source build steps have been added as part of this update.

* Add CMake to the ninja build

* Update Dockerfile to fix ARM64 build issue

The Dockerfile for Centos7 was updated to better support ARM64 architectures. Instead of fetching a precompiled binary of CMake, we now clone CMake v3.28.1 and compile it specifically for the target. This change includes the addition of the `-mno-outline-atomics` flag, which resolves ARM64-specific issues with compilation.

* Update Dockerfile to conditionally set CFLAGS for arm64 architecture

The Dockerfile for CentOS7 now includes a condition that sets the CFLAGS environment for arm64 architecture. This change was made during the process of enabling CMake to bootstrap itself and ensures that correct flags are set for specific architectures during the build process.