Move TokenLenBytes to defaults package (#37279)
Some consts in the `auth` package are used in a lot of other packages.
This increases the probality of getting a cyclic dependency error.

This PR moves some of those consts to the defaults package and updates
all the required locations of their usage.

Only one is missing in the `/e/` repo, which I'm going to open a PR for
it.
After merging, we can also remove the `lib/auth.TokenLenBytes`