Update to BoringSSL `fips-20220613` tag (#36166)
* Add ninja-build setup to Dockerfile-centos7

The Dockerfile-centos7 has been updated to include the setup and installation of ninja-build. The version of ninja-build included in CentOS 7 AMR64 was too old for building boringssl, hence it needs to be built from source. Both additional dependencies and the source build steps have been added as part of this update.

* Add CMake to the ninja build

* Update Dockerfile to fix ARM64 build issue

The Dockerfile for Centos7 was updated to better support ARM64 architectures. Instead of fetching a precompiled binary of CMake, we now clone CMake v3.28.1 and compile it specifically for the target. This change includes the addition of the `-mno-outline-atomics` flag, which resolves ARM64-specific issues with compilation.

* Update Dockerfile to conditionally set CFLAGS for arm64 architecture

The Dockerfile for CentOS7 now includes a condition that sets the CFLAGS environment for arm64 architecture. This change was made during the process of enabling CMake to bootstrap itself and ensures that correct flags are set for specific architectures during the build process.

* Update Clang from 12 to 14 to match BoringSSL `fips-20220613` tag

Matches what upstream Go is using for BoringCrypto build:
https://github.com/golang/go/commit/7383b2a4db5dc93c9b875b42d5add73d27cc4b9f

Also set `CMAKE=cmake3`, as `boring-rs` (by way of `cmake-rs`) calls `cmake`.

Additionally, make `build-centos7-assets` a dependency.

* Removes build-centos7-assets as a dependency for the build-centos7 target.
Adds a comment to the Makefile to explain why.

* Updated Clang build configuration in Dockerfile

Updates clang to be built with the inclusion of the header files
necessary for building boring. Also changes the build to use Ninja.

* moves the ninja copy closer to the clang copy

* Swap to fork of `boring-rs` for build fixes (#36720)

BoringSSL does not build properly on CentOS 7 without some build fixes. As these fixes are not yet upstreamed, a separate fork is being used -- https://github.com/gravitational/boring.

Co-authored-by: Isaiah Becker-Mayer <isaiah@goteleport.com>

* Ninja build snuck back into Dockerfile-centos7, but we don't need it since it's in Dockerfile-centos7-assets now. Removing it.

* Prevent docker buildkit from caching

Use docker build to build fips image to prevent docker from fetching the base centos image from the registry.

* Add 'nolint' comment in boringtest.go

A 'nolint' comment was added in the boringtest go file to ignore linting this file

* Rename test Go file

---------

Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>
Co-authored-by: Isaiah Becker-Mayer <isaiah@goteleport.com>