Add linters for consistent docs product language
Closes #19214

Maintaining a consistent language for talking about Teleport products is
important for ensuring clarity in the documentation. This change adds
rules to our Vale linting config for docs to help maintain a consistent
product language:

- **Teleport account URL examples:** Require `example.teleport.sh`
  unless the text is mentioning the status page.
- **Service names:** Require these to be capitalized.
- **Enforce consistency in other areas where we often get
  discrepancies:** e.g., "Auth Service instances" instead of "auth
  servers".

This change also renames the `messaging` style rule to
`consistent-terms` so we don't have a `messaging.messaging` rule.

Note that these linters overlap with similar ones that we run from the
`gravitational/docs` repo. The plan is to remove the latter once we
merge the linters here.