AWS OIDC: remove config dependency in ExternalAuditStorage (#37315)
* AWS OIDC: remove config dependency in ExternalAuditStorage

Currently the ExternalAuditStorage methods require the `config` package
to re-use the configuration struct used in `teleport integration configure`

We are refactoring the AWS OIDC actions to use a gRPC Service.
Importing the `config` package makes this refactoring a bit difficult
because of a cycling dependency: `awsoidc` requires `config` which
requires `auth` which will require the `awsoidc service` which requires
the `awsoidc` package.

This PR removes the usage of the `config` package in `awsoidc` package.

* rename type and godoc

* move type to servicecfg

* fix gci