ci: Ensure pinned versions when installing wasm-pack (#3308)
When installing `wasm-pack` with `cargo`, use the `--locked` flag to
ensure the versions referenced in the Cargo.lock file are used and not
any newer versions. This creates a repeatable build and insulates us
from upstream breakage in new versions.

In particular an upstream version change included the hoot-0.1.2
dependency on a new build which fails to build with rust 1.71.1 , which
is the version we are using.

Link: https://github.com/rustwasm/wasm-pack/issues/1365

Source-commit: gravitational/teleport.e@c7391c00c7573bf91707c4c7625a7c57630b507e
Source-pull-requests: gravitational/teleport.e#3308