Core SCIM backend (#3203)
Implements the back-end of a GRPC SCIMService, based loosely on the example HTTP service
in the elimity SCIM library (See: https://github.com/elimity-com/scim).

Unfortunately, I wasn't able to use the library HTTP implementation directy, mainly
because:

 * The SCIM implementation is tightly bound to being inside a `http.Handler`, and we
   needed to use gRPC to forward requests from the proxy to the auth. The SCIM handling
   needs to be done on auth, as this is the only place where we can read plugin
   credentials.

 * Okta's request bodies are not always schema-compliant, and the elimity library had
   baked-in schema validation.

 * Our SCIM service needed another dimension of freedom. The Elimity library abstracts
   over the concrete User and Group types to provide a uniform structure for dealing
   with any kind of resource.

On top of that, our SCIM implementation needs to be able to handle the idiosyncratic
behaviour of multiple upstream IdPs, and this wasn't flexibility the library could give
us.

In this implemetation,

 * the top-level `Service` handles incoming GRPC requests for all resource types
 * `resourceTypeHandlers` handle all of the resource-type-specific SCIM behaviour that
   can be customised by simple configuration (e.g. ensuring schema attributes are set,
   etc).
 * `resourceHandler` implementations encapsulate Resource-type-specific behaviours
   that actually need different code paths (e.g. Users need to unpack a SCIM resource
   as a User, and perform CRUD operations on Teleport User resources, whereas Groups
   have entirely different handling requirements)

Cutting across all of the above is the `providerShim`, which allows the SCIM service to
handle the idiosyncratic behaviour of multiple upstream IdPs (and how Teleport deals
with them). The providerShim handles IdP-specific tasks like converting SCIM resources
to/from Teleport Resources, authentication, and so on.

Derived-From: #3152

Source-commit: gravitational/teleport.e@03cbfea117d410d3a1329a77b27b1a625493830d
Source-pull-requests: gravitational/teleport.e#3203